+++ CHANGELOG.md
@@ -38,6 +38,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   and preventing memory allocations in the process.
 - Further optimized VST3 audio processing by preallocating small vectors for
   event and parameter change queues.
+- VST2 audio processing also received the same small vector optimization to get
+  rid of any last potential allocations during audio processing.
 - Changed the way mutual recursion in VST3 plugins on the plugin side works to
   counter any potential GUI related timing issues with VST3 plugins.
 - The deserialization part of yabridge's communication is now slightly faster by